+++ CHANGELOG.md
@@ -57,6 +57,8 @@ See also https://github.com/neo4j/neo4j-python-driver/wiki for a full changelog.
 - Remove deprecated class `neo4j.Bookmark` in favor of `neo4j.Bookmarks`.
 - Remove deprecated class `session.last_bookmark()` in favor of `last_bookmarks()`.
 - Make undocumented classes `ResolvedAddress`, `ResolvedIPv4Address`, and `ResolvedIPv6Address` private.
+- No longer implicitly closing drivers and sessions in `__del__()` (finalizer/destructor).
+  Make sure to call `.close()` on them explicitly or use them in a `with` statement.
 
 
 ## Version 5.28